+++ src/__playground__/index.ts
@@ -1,10 +1,3 @@
-import { Messenger } from '@messenger';
+import { worker } from './worker.js';
 
-const m1 = new Messenger();
-const m2 = new Messenger(m1.raw);
-
-m1.closeAll();
-
-await new Promise((r) => setTimeout(r, 5e3));
-
-console.log(m2.closed, m1.closed);
+await worker({ name: 'wait', reffed: false, shareEnv: false });